requirements.txt
yfinance
pandas
PyYAML
python-dotenv
nsetools
requests
pytz
